+++ src/proto_alpha/lib_client_commands/client_proto_context_commands.ml
@@ -778,7 +778,8 @@ let commands_ro () =
             "%a"
             Data_encoding.Binary_schema.pp
             (Data_encoding.Binary.describe
-               Alpha_context.Operation.unsigned_encoding)
+               Alpha_context.Operation
+               .unsigned_encoding_with_legacy_attestation_name)
         in
         return_unit);
     command