+++ src/proto_alpha/lib_protocol/operation_repr.mli
@@ -530,13 +530,35 @@ val manager_kind : 'kind manager_operation -> 'kind Kind.manager
 
 val encoding : packed_operation Data_encoding.t
 
+(** Operation encoding that accepts legacy attestation name : `endorsement`
+    (and preendorsement, double_<op>_evidence) in JSON
+
+    https://gitlab.com/tezos/tezos/-/issues/5529
+
+    This encoding is temporary and should be removed when the endorsements kinds
+    in JSON will not be accepted any more by the protocol.
+*)
+val encoding_with_legacy_attestation_name : packed_operation Data_encoding.t
+
 val contents_encoding : packed_contents Data_encoding.t
+
+val contents_encoding_with_legacy_attestation_name :
+  packed_contents Data_encoding.t
 
 val contents_list_encoding : packed_contents_list Data_encoding.t
 
+val contents_list_encoding_with_legacy_attestation_name :
+  packed_contents_list Data_encoding.t
+
 val protocol_data_encoding : packed_protocol_data Data_encoding.t
 
+val protocol_data_encoding_with_legacy_attestation_name :
+  packed_protocol_data Data_encoding.t
+
 val unsigned_operation_encoding :
+  (Operation.shell_header * packed_contents_list) Data_encoding.t
+
+val unsigned_operation_encoding_with_legacy_attestation_name :
   (Operation.shell_header * packed_contents_list) Data_encoding.t
 
 val raw : _ operation -> raw