+++ src/proto_alpha/lib_sc_rollup_node/configuration.ml
@@ -208,7 +208,7 @@ let protocol_max_batch_size =
   in
   Protocol.Constants_repr.max_operation_data_length
   - Data_encoding.Binary.length
-      Operation.encoding
+      Operation.encoding_with_legacy_attestation_name
       (Operation.pack empty_message_op)
 
 let default_batcher_max_batch_size = protocol_max_batch_size